# Remove containers from deleted pods once containers have exited