@dashpole 2017-10-02
I tested this PR by doing the following:
I disabled periodic container gc by commenting out that code (just for my testing).  I did this to ensure this PR is what is causing containers to be cleaned up.
I ran a node-e2e test which creates and deletes a pod.

I checked the kubelet log during the test, and found that the pod is stopped, cleaned up, and removed within 2 seconds of observing the deletion timestamp:
`16:55:59 SyncLoop (DELETE ...`
`16:56:00 GenericPLEG ... running -> exited`
`16:56:00 Removing container ...`
`16:56:01 Volume detached for volume ...`
`16:56:01 genericPLEG ... exited -> non-existent`
`16:56:01 ... fully terminated and removed from etcd`