@smarterclayton 2017-10-02
I wish we had a metric that reported end to end pod deletion metrics