@smarterclayton 2017-10-02
Started the kubemark 500 run to test