@shyamjvs 2017-10-02
@smarterclayton I think running this against kubemark wouldn't give us much signal because there's no real containers to be deleted there - with a mocked kubelet-gc that pretty much deletes the pods instantaneously (ref https://github.com/kubernetes/kubernetes/issues/51899#issuecomment-332887709). Besides, I'm also seeing that the `delete pods` latency for kubemark isn't much different for:

this PR

```
I1002 19:01:45.811] Oct  2 19:01:45.810: INFO: Top latency metric: {Resource:pods Subresource: Verb:DELETE Scope:namespace Latency:{Perc50:3.522ms Perc90:6.248ms Perc99:12.048ms Perc100:0s} Count:31000}
```

vs

[CI run](https://storage.googleapis.com/kubernetes-jenkins/logs/ci-kubernetes-kubemark-500-gce/8818/artifacts/APIResponsiveness_density_2017-10-02T18:46:18Z.json) against head:

```
    {
      "data": {
        "Perc50": 3.679,
        "Perc90": 6.915,
        "Perc99": 12.046
      },
      "unit": "ms",
      "labels": {
        "Count": "31000",
        "Resource": "pods",
        "Scope": "namespace",
        "Subresource": "",
        "Verb": "DELETE"
      }
    },
```

I'll try testing against real 2k cluster tomorrow if that SGTY.