@shyamjvs 2017-10-03
Seems to help quite a bit:

without the PR:

```
Oct  3 17:14:21.822: INFO: WARNING Top latency metric: {Resource:pods Subresource: Verb:DELETE Scope:namespace Latency:{Perc50:4.445ms Perc90:612.189ms Perc99:5.041177s Perc100:0s} Count:124002}
```

with the PR:

```
Oct  3 14:12:06.169: INFO: Top latency metric: {Resource:pods Subresource: Verb:DELETE Scope:namespace Latency:{Perc50:3.77ms Perc90:8.846ms Perc99:426.462ms Perc100:0s} Count:124000}
```