@shyamjvs 2017-10-03
@dashpole @dchen1107 Can we have this PR in soon'ish? Need to get the scale tests back to green.