@smarterclayton 2017-10-03
No further comments from me?