@wojtek-t 2017-10-03
Once merged we should cherrypick it to 1.8