@liggitt 2017-10-04
pick to 1.8 opened at https://github.com/kubernetes/kubernetes/pull/53422